scripts/requirements.txt
PyYAML>=6.0.1,<7
python-dotenv>=1.0.1,<2
colorama>=0.4.6
jsonschema>=4.23.0,<5
